enforce default pre-render;regroup mods;simplify `use`
