+++ src/tests/mod.rs
@@ -1,3 +1,19 @@
+use super::*;
+
+#[test]
+fn test_name() {
+    let pre = KatexProcessor;
+    let preprocessor: &dyn Preprocessor = &pre;
+    assert_eq!(preprocessor.name(), "katex")
+}
+
+#[test]
+fn test_support_html() {
+    let preprocessor = KatexProcessor;
+    assert!(preprocessor.supports_renderer("html"));
+    assert!(preprocessor.supports_renderer("other_renderer"))
+}
+
 mod escape_test;
 
 #[cfg(feature = "pre-render")]
